requirements.txt
flask
flask-cors
langchain
langchain-openai
langchain-groq
langchain-community
langsmith
sqlalchemy
psycopg2
python-dotenv
psycopg2-binary
gunicorn